requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
twilio==9.4.3
anthropic==0.49.0
httpx==0.27.2
